% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviatedPlatforms.R
\docType{data}
\name{abbreviatedPlatforms}
\alias{HM27}
\alias{HM450}
\alias{HM850}
\alias{HMEPIC}
\alias{HUEX1}
\alias{PLUS2}
\alias{abbreviatedPlatforms}
\alias{platforms}
\alias{shortcuts}
\title{quick shortcuts for some platforms; could certainly use more
FIXME: add something similar for SRAdb, eg. BISULFITE, RNASEQ, ATACSEQ, etc.}
\format{An object of class \code{character} of length 1.}
\usage{
getRecent(HUEX1)

getRecent(PLUS2)

getRecent(HM27)

getRecent(HM450)

getRecent(HMEPIC)

getRecent(HM850)
}
\description{
quick shortcuts for some platforms; could certainly use more
FIXME: add something similar for SRAdb, eg. BISULFITE, RNASEQ, ATACSEQ, etc.

HUEX1 -- Affy HuEx array

PLUS2 -- Affy hgu133plus2 array

HM27 -- Illumina HumanMethylation27 array

HM450 -- Illumina HumanMethylation450 array

HMEPIC -- Illumina HumanMethylationEPIC array

HM850 -- Illumina HumanMethylationEPIC array
}
\seealso{
\code{\link{getRecent}}, \code{\link{getDetails}}, \code{\link{getExperiments}}, \code{\link{getGSE}}, \code{\link{getSRA}}, and so forth
}
\keyword{datasets}

