% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForUpdates.R
\name{checkForUpdates}
\alias{checkForUpdates}
\title{Download an up-to-date GEOmetadb SQLite file if it has been more than a day
FIXME: look at the actual filename to determine whether it's been updated!}
\usage{
checkForUpdates(whatKind = c("GEO", "SRA"))
}
\arguments{
\item{whatKind}{either "GEO" or "SRA" (the latter is not well supported yet)}
}
\description{
Download an up-to-date GEOmetadb SQLite file if it has been more than a day
FIXME: look at the actual filename to determine whether it's been updated!
}

