% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDetails.R
\name{getDetails}
\alias{gD}
\alias{getDetails}
\title{get details for a specific series (experiment)}
\usage{
getDetails(gse)

gD(gse)
}
\arguments{
\item{gse}{the series identifier}
}
\value{
a short list of details
}
\description{
get details for a specific series (experiment)
}
\section{Functions}{
\itemize{
\item \code{gD}: 
}}

