% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{gE}
\alias{getExperiments}
\title{get the most recently updated experiments for a given search phrase}
\usage{
getExperiments(string, limit = 10, gpl = NULL)

gE(string, limit = 10, gpl = NULL)
}
\arguments{
\item{string}{the phrase to search for}

\item{limit}{optional, max records to return (default: 10)}

\item{gpl}{optional, specific platform to limit to (default: NULL)}
}
\value{
a data.frame, containing between 0 and ((from-to) + 1) rows
}
\description{
get the most recently updated experiments for a given search phrase
}
\section{Functions}{
\itemize{
\item \code{gE}: two-letter shortcut for lazy typists
}}

