% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFASTQ.R
\name{getFASTQ}
\alias{gFQ}
\alias{getFASTQ}
\title{get FASTQs for an SRA series (SRP)}
\usage{
getFASTQ(SRP, ...)
}
\arguments{
\item{SRP}{the series identifier}

\item{...}{optional arguments, e.g. destDir}
}
\description{
get FASTQs for an SRA series (SRP)
}

