% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRA.R
\name{getSRA}
\alias{gSRA}
\alias{getSRA}
\title{get SRA entries, minus hassling with the database downloads}
\usage{
getSRA(...)

gSRA(...)
}
\arguments{
\item{...}{arguments for SRAdb::getSRA (besides sra_con, which is automated)}
}
\description{
get SRA entries, minus hassling with the database downloads
}
\section{Functions}{
\itemize{
\item \code{gSRA}: shortcut for lazy typists
}}

