% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRP.R
\name{getSRP}
\alias{gSRP}
\alias{getSRP}
\title{get samples (or other stuff) for an SRA series (SRP) or other identifier}
\usage{
getSRP(SRP, ...)

gSRP(SRP, ...)
}
\arguments{
\item{SRP}{the series identifier}

\item{...}{other arguments to pass to SRAdb::sraConvert}
}
\value{
the samples and some descriptions
}
\description{
get samples (or other stuff) for an SRA series (SRP) or other identifier
}
\section{Functions}{
\itemize{
\item \code{gSRP}: shortcut for lazy typists
}}

