% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRPfromGSE.R
\name{getSRPfromGSE}
\alias{GSE2SRP}
\alias{getSRPfromGSE}
\title{try to extract the associated SRP study from a GEO GSE, if one exists}
\usage{
getSRPfromGSE(GSE)

GSE2SRP(GSE)
}
\arguments{
\item{GSE}{the GSE}
}
\value{
the SRP or SRPs, if found
}
\description{
try to extract the associated SRP study from a GEO GSE, if one exists
}
\section{Functions}{
\itemize{
\item \code{GSE2SRP}: shortcut for lazy typists
}}

