% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{gS}
\alias{getSummary}
\title{get a summary for a specific series (experiment)}
\usage{
getSummary(gse, collapse = "; ")
}
\arguments{
\item{gse}{the series identifier}

\item{collapse}{optional string to use for collapsing fields ("; ")}
}
\value{
a string with the collapsed details
}
\description{
get a summary for a specific series (experiment)
}

