% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strpop.R
\name{strpop}
\alias{strpop}
\title{a convenience function to pop pieces off of a string}
\usage{
strpop(x, y = " ", z = NULL)
}
\arguments{
\item{x}{the string}

\item{y}{the split}

\item{z}{(optional) post-split elements to return (default: the last one)}
}
\value{
a character vector, possibly of length greater than one
}
\description{
this should really just go in timlib or whatever the hell I end up calling it
}

