% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDetails.R, R/getSummary.R
\name{getDetails}
\alias{gD}
\alias{gS}
\alias{getDetails}
\title{get details for a specific series (experiment)}
\usage{
getDetails(gse)

gD(gse)

gS(gse, collapse = "; ")
}
\arguments{
\item{gse}{the series identifier}
}
\value{
a short list of details
}
\description{
get details for a specific series (experiment)
}
\section{Functions}{
\itemize{
\item \code{gD}: two-letter shortcut for lazy typists

\item \code{gS}: two-letter shortcut for lazy typists
}}

