% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSE.R
\name{getGSE}
\alias{gG}
\alias{getGSE}
\title{get SummarizedExperiment(s) for a GSE(s) (usually expression, maybe DNA meth)}
\usage{
getGSE(gse, combine = FALSE, map = FALSE)

gG(gse, combine = FALSE, map = FALSE)
}
\arguments{
\item{gse}{the series identifier}

\item{combine}{try to flatten (cbind) the resulting SEs? (default: FALSE)}

\item{map}{try to map the rows to sensible symbols? (default: FALSE)}
}
\value{
a list of the SE(s) or subclass(es) thereof
}
\description{
get SummarizedExperiment(s) for a GSE(s) (usually expression, maybe DNA meth)
}
\section{Functions}{
\itemize{
\item \code{gG}: shortcut for lazy typists
}}

