% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFASTQ.R, R/getSRP.R
\name{gFQ}
\alias{gFQ}
\alias{gSRP}
\alias{getSRP}
\title{get samples for an SRA series (SRP)}
\usage{
gFQ(SRP, ...)

getSRP(SRP)

gSRP(SRP)
}
\arguments{
\item{SRP}{the series identifier}
}
\value{
the samples and some descriptions
}
\description{
get samples for an SRA series (SRP)
}
\section{Functions}{
\itemize{
\item \code{gFQ}: shortcut for lazy typists

\item \code{gSRP}: shortcut for lazy typists
}}

